# Improve pathfinding performance by using a heap to store traversable polygons