## Description

Improves pathfinding performance by doing 2 things.
- Makes retrieval of already visited polygons constant time instead of linear time
- Uses a heap to store the polygons to visit instead of a linked list
  - To implement this, introduces a generic heap structure that notifies of element index changes and also allows to update an arbitrary element's position (the existing functions in `SortArray` don't allow for that)

Closes https://github.com/godotengine/godot-proposals/issues/8496.

### Changes in time complexity

`n`: number of polygons visited until now (`navigation_polys`)
`m`: number of polygons marked for later visiting (`traversable_polys`)

In every scenario, `m` increases at a much slower rate than `n`.

- An individual neighbor polygon check/registration/update goes from `O(n)` to `O(log2(m))` (or `O(1)` if the polygon has already been visited and no update is needed)
- Retrieving the next polygon to visit goes from `O(m)` to `O(log2(m))`

These changes should result in a big performance improvement on big maps.

## Test project

[Navigation path finding tests.zip](https://github.com/godotengine/godot/files/13800063/Navigation.path.finding.tests.zip)

## Benchmark numbers

Below are execution times for `NavigationServer*D.map_get_path` before and after this PR on different map sizes.

Map size | Iteration count | Average time - master | Average time - PR | Speed-up
--- | --- | --- | --- | ---
15x10 cells (small) | 1000 | 0.047ms | 0.026ms | x1.80 faster
90x50 cells (normal) | 500 | 5.976ms | 0.688ms | x8.68 faster
240x160 cells (big) | 10 | 465.4ms | 11.3ms | x41.1 faster

As you can see, the bigger the map, the bigger the speed-up.

- The measurements were done on a build made with `scons platform=windows` (no debug info)
- Done on Windows 10.0.19045 with an AMD Ryzen 7 2700X (4.3 GHz max, 8 core, 16 threads)
- You can find the maps in the test project linked above

## Further improvements

There is room for further improvement here.
- Right now memory is allocated on the heap on each path request. It's rather wasteful but doesn't need to. We could keep a pool of arrays that would be reused each request.
- Finding the start and end polygon before the actual pathfinding is currently an `O(n)` operation. We will invariably loop over every polygon each request. We could use a specialized data structure to improve performance here.
- With the current API, even if we're interested only in paths that actually reach the target, a path to the closest reachable polygon will always be calculated if the target is unreachable. That's two pathfinding requests even if we only want one. We could add a new option/method to cover this use case. That one requires a proposal though.

I intend to work on these if there is no problem with this PR.