@ershn 2023-12-29
I realized that the current implementation doesn't take care of multithreading.
Trying to find a path from multiple threads on the same map at the same time will currently cause a crash.

I'm turning this PR back into a draft until I fix the problem.