@ershn 2023-12-31
I changed the approach to be multithreading friendly. Pathfinding requests now leave `Polygon` objects untouched.

1. On map construction (`NavMap::sync`), assign each `Polygon` an id which is its index in the vector
2. On entering `map_get_path`, create a `NavigationPoly` vector of the same size as the one above
3. Use the `Polygon` id to find the corresponding `NavigationPoly` in the vector and vice versa

Since it's not possible anymore for the `NavigationPoly` vector to grow in size during pathfinding, I also changed it so that I now store pointers in the heap instead of vector indexes. It improved readability and also slightly increased performance.

I added a multithreading test to the test project and also rebased on master (checked that https://github.com/godotengine/godot/issues/85238 is still fixed).