@smix8 2024-04-03
I suggest moving this to a navigation module internal class/struct instead of a core template.

I talked with some core devs and there is a preference for a more "local" solution that does not touch core.