@ershn 2024-08-29
Sorry for the delay. I moved the heap class to `nav_utils.h`, rebased on master, squashed the commits and checked that nothing broke after that.
I wasn't entirely sure where to put the new unit test file so feel free to change that if necessary.

Thank you for the review smix8 !