@smix8 2024-08-29
> I wasn't entirely sure where to put the new unit test file so feel free to change that if necessary.

@Scony is usually the test maniac ... ;)
... but I would say those tests would belong to the `tests/servers/test_navigation_server_3d.h` file.