@ershn 2024-08-31
Thank you for the tests !

The bug probably happened because the synthetic polygons generated for navigation links were not being taken care of during map sync. Their id was never updated from the default value of UINT32_MAX which resulted in invalid array accesses when trying to process them during pathfinding.

This should be fixed now. Could you test again with your test scenes ?

> but you can imo also just leave it in that separated file as is for now
Ok, I'll leave the tests there for the time being then.