@smix8 2024-08-31
> This should be fixed now. Could you test again with your test scenes ?

@ershn  Yeah that was it, no more crashes or errors (so far 😝)