@smix8 2024-09-01
> In my view, the heap implementation is generic enough so that should placed be within `core/templates`. @smix8 who are the "core devs" that suggested otherwise?

That was when I asked reduz with 1-2 other core people around for a review on this PR months ago. They all said this should be local to the navigation module and not part of core which is why I made that comment back in April to unblock this PR.