@kleonc 2024-09-01
> In my view, the heap implementation is generic enough so that should placed be within `core/templates`.

@Scony See [this](https://docs.godotengine.org/en/4.3/contributing/development/best_practices_for_engine_contributors.html#prefer-local-solutions). Currently it's used only by the navigation so adding to core is not justified. If something else will need it as well then it will probably be fine to move it to core.
(I'm not a core team member but IIRC that's the reasoning.)